07/03/23

+++ Java/ExerciciosExtraIfElser/NumeroPar.java
@@ -4,10 +4,10 @@ import java.util.Scanner;
 
 public class NumeroPar {
     Scanner Leitura = new Scanner(System.in);
-    double numero;
+    int numero;
     public void par(){
         System.out.println("Digite um numero para saber se ele é impar ou par: ");
-        numero = Leitura.nextDouble();
+        numero = Leitura.nextInt();
         if (numero % 2 == 0){
             System.out.println(numero +" é PAR!");
         } else {
